♻️ Organize models
